Fix pwrup_delay result type.

+++ sys/dev/bhnd/cores/chipc/pwrctl/bhnd_pwrctl_private.h
@@ -40,7 +40,7 @@ int		bhnd_pwrctl_init(struct bhnd_pwrctl_softc *sc);
 int		bhnd_pwrctl_setclk(struct bhnd_pwrctl_softc *sc,
 		    bhnd_clock clock);
 uint32_t	bhnd_pwrctl_getclk_speed(struct bhnd_pwrctl_softc *sc);
-uint16_t	bhnd_pwrctl_fast_pwrup_delay(struct bhnd_pwrctl_softc *sc);
+u_int		bhnd_pwrctl_fast_pwrup_delay(struct bhnd_pwrctl_softc *sc);
 
 /**
  * If supported by the chipset, return the clock source for the given clock.
